Fix JC-556: 1. Add color to active link in 'header.component' 2. Move 'poster-popup' to 'components' 3. Update params and 'animeDetails$' in 'details.component' 4. Delete 'navigate' from 'logout' in components. 5. Add 'usLoggedIn$' subscriber to 'app.component'
