+++ apps/angular/src/app/app.component.ts
@@ -1,4 +1,7 @@
-import { Component } from '@angular/core';
+import { Component, inject } from '@angular/core';
+import { Router } from '@angular/router';
+
+import { AuthService } from '../core/services/auth.service';
 
 /** App component. */
 @Component({
@@ -6,4 +9,16 @@ import { Component } from '@angular/core';
 	templateUrl: './app.component.html',
 	styleUrls: ['./app.component.css'],
 })
-export class AppComponent {}
+export class AppComponent {
+	private readonly authService = inject(AuthService);
+
+	private readonly router = inject(Router);
+
+	public constructor() {
+		this.authService.isLoggedIn$.subscribe(isLoggedIn => {
+			if (isLoggedIn === false) {
+				this.router.navigate(['/anime/table']);
+			}
+		});
+	}
+}
