+++ apps/angular/src/app/features/anime/components/header/header.component.html
@@ -1,5 +1,6 @@
 <header class="header">
 	<a *ngIf="isLoggedIn$ | async" routerLink="/home" mat-raised-button>Home Page</a>
+	<a *ngIf="(isLoggedIn$ | async)" routerLink="/anime/table" routerLinkActive="active-link" mat-raised-button>Anime table</a>
 	<a *ngIf="(isLoggedIn$ | async) === false" routerLink="/home" mat-raised-button>Login</a>
-	<button *ngIf="isLoggedIn$ | async" type="button" routerLink="/anime/table" mat-raised-button color="primary" (click)="logout()">Log out</button>
+	<button *ngIf="isLoggedIn$ | async" type="button" mat-raised-button color="primary" (click)="logout()">Log out</button>
 </header>
