+++ apps/angular/src/app/features/anime/components/poster-popup/poster-popup.component.html
