+++ apps/angular/src/app/features/anime/details/details.component.ts
@@ -1,11 +1,11 @@
 import { Dialog } from '@angular/cdk/dialog';
 import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
+import { ActivatedRoute } from '@angular/router';
 import { AnimeService } from '@js-camp/angular/core/services/anime.service';
-import { AuthService } from '@js-camp/angular/core/services/auth.service';
 import { AnimePoster } from '@js-camp/core/models/anime-poster';
+import { switchMap } from 'rxjs';
 
-import { PosterPopupComponent } from './poster-popup/poster-popup.component';
+import { PosterPopupComponent } from '../components/poster-popup/poster-popup.component';
 
 /** Details of anime. */
 @Component({
@@ -18,29 +18,14 @@ export class AnimeDetailsComponent {
 
 	private readonly route = inject(ActivatedRoute);
 
-	private readonly router = inject(Router);
-
-	private readonly authService = inject(AuthService);
-
 	private readonly animeService = inject(AnimeService);
 
-	private readonly animeId = this.route.snapshot.params['id'];
+	/** List of anime. */
+	protected readonly animeDetails$ = this.route.params.pipe(
+		switchMap(params => this.animeService.getAnimeDetails(params['id'])),
+	);
 
 	private readonly dialog = inject(Dialog);
-
-	/** List of anime. */
-	protected readonly animeDetails$ = this.animeService.getAnimeDetails(this.animeId);
-
-	private getIdFromParams(): number {
-		const { id } = this.route.snapshot.params;
-		return id;
-	}
-
-	/** Log out and redirect to anime. */
-	protected logout(): void {
-		this.authService.logout();
-		this.router.navigate(['/anime/table']);
-	}
 
 	/**
 	 * Opens poster popup.
