+++ apps/angular/src/app/features/auth/profile/profile.component.ts
@@ -1,5 +1,4 @@
 import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
-import { Router } from '@angular/router';
 import { AuthService } from '@js-camp/angular/core/services/auth.service';
 
 /** User profile. */
@@ -12,11 +11,8 @@ import { AuthService } from '@js-camp/angular/core/services/auth.service';
 export class ProfileComponent {
 	private readonly authService = inject(AuthService);
 
-	private readonly router = inject(Router);
-
 	/** Log out and redirect to anime. */
 	protected logout(): void {
 		this.authService.logout();
-		this.router.navigate(['/anime/table']);
 	}
 }
