+++ apps/angular/src/core/services/logged-auth-guard.service.ts
@@ -10,8 +10,6 @@ import { AuthService } from './auth.service';
 })
 export class LoggedAuthGuard implements CanActivate {
 
-	private readonly auth = inject(AuthService);
-
 	private readonly router = inject(Router);
 
 	private readonly authService = inject(AuthService);
